requirements.txt
pygame
pyaudio
SpeechRecognition
vosk
gTTS
langchain==0.0.142
openai
python-dotenv
colorama
retry
sounddevice